GPIO/raspberry-pi5-gpio-test/requirements.txt
RPi.GPIO
keyboard
```